🚚 Rename: useQuery 훅 queries 폴더 하위로 이동
