+++ src/components/auth/loginForm/LoginForm.tsx
@@ -8,7 +8,7 @@ import useShowPassword from '@/hooks/useShowPassword'
 import { useCookies } from 'react-cookie'
 import { useEffect, useState } from 'react'
 import Button from '@/components/common/button/Button'
-import useLogin from '@/hooks/useLogin'
+import useLogin from '@/hooks/queries/useLogin'
 import { LoginDataType } from '@/types/auth/loginDataType'
 import useModal from '@/hooks/useModal'
 import { usePathname, useRouter } from 'next/navigation'
