+++ src/components/auth/signupForm/SignupForm.tsx
@@ -12,7 +12,7 @@ import { useRouter } from 'next/navigation'
 import useModal from '@/hooks/useModal'
 import useCheckDuplicateNickname from '@/hooks/useCheckDuplicateNickname'
 import { SignupDataType } from '@/types/auth/signupDataType'
-import useLogin from '@/hooks/useLogin'
+import useLogin from '@/hooks/queries/useLogin'
 import { signup } from '@/apis/auth/signup'
 import { emailSendVerify } from '@/apis/auth/emailSendVerify'
 import { emailCheckVerify } from '@/apis/auth/emailCheckVerify'
